# api/openapi.yaml
openapi: 3.1.0
info:
  title: YORVA Local Node API
  version: 0.0.0
  description: Local loopback contract between YORVA Desktop and yorvad through Phase 4 instance inventory.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "49152"
        description: Ephemeral loopback port returned by the bootstrap handshake.
...
paths:
...
  /api/v1/node:
    get:
      operationId: getNode
      summary: Read the persisted local Node identity
      tags: [Node]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The local Node identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsNode
      summary: Validate CORS access for the Node endpoint
      tags: [Node]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/events:
    get:
      operationId: streamEvents
      summary: Connect to the authenticated local event stream
      tags: [Bootstrap]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An SSE stream of redacted notifications, including operation.started, operation.progress, operation.completed, operation.failed and operation.cancelled. Resource GET remains the source of truth.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsEvents
      summary: Validate CORS access for the event stream endpoint
      tags: [Bootstrap]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runtimes/{runtimeKind}/detect:
    parameters:
      - name: runtimeKind
        in: path
        required: true
        schema:
          type: string
          enum: [hermes]
    post:
      operationId: detectRuntime
      summary: Discover a registered Runtime installation
      description: Performs a bounded, read-only discovery. Negative discovery outcomes are returned as typed 200 responses.
      tags: [Runtimes]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The completed Runtime discovery result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeDiscovery"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
    options:
      operationId: optionsRuntimeDetection
      summary: Validate CORS access for Runtime discovery
      tags: [Runtimes]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runtimes/hermes/install:
    post:
      operationId: startHermesInstall
      summary: Start a Hermes installation Operation
      tags: [Runtimes]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeInstallRequest"
      responses:
        "202":
          description: The installation Operation was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsHermesInstall
      summary: Validate CORS access for Hermes installation
      tags: [Runtimes]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runtimes/hermes/prerequisites:
    get:
      operationId: getHermesPrerequisites
      summary: Read Hermes Node.js prerequisite health
      tags: [Runtimes]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Typed Node/npm/dependency health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HermesPrerequisites"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    options:
      operationId: optionsHermesPrerequisites
      summary: Validate CORS access for Hermes prerequisites
      tags: [Runtimes]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runtimes/hermes/prerequisites/install:
    post:
      operationId: startHermesPrerequisites
      summary: Start a Hermes Node prerequisite Operation
      tags: [Runtimes]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeInstallRequest"
      responses:
        "202":
          description: The prerequisite Operation was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    options:
      operationId: optionsHermesPrerequisitesInstall
      summary: Validate CORS access for Hermes prerequisite installation
      tags: [Runtimes]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/operations:
    get:
      operationId: listOperations
      summary: List recent Operations for Desktop recovery
      tags: [Operations]
      security:
        - bearerAuth: []
      parameters:
        - name: targetType
          in: query
          schema:
            type: string
        - name: targetId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: A bounded Operation list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsOperations
      summary: Validate CORS access for Operation listing
      tags: [Operations]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/operations/{operationId}:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getOperation
      summary: Read one Operation
      tags: [Operations]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The Operation resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsOperation
      summary: Validate CORS access for an Operation
      tags: [Operations]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/operations/{operationId}/log:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getOperationLog
      summary: Read diagnostic install log text for one Operation
      tags: [Operations]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Temporary diagnostic log text for this Operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsOperationLog
      summary: Validate CORS access for Operation diagnostic log
      tags: [Operations]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/operations/{operationId}/cancel:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: cancelOperation
      summary: Cancel a non-terminal Operation
      tags: [Operations]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The Operation after cancellation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsCancelOperation
      summary: Validate CORS access for Operation cancellation
      tags: [Operations]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runtimes/{runtimeId}/instances:
    parameters:
      - name: runtimeId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listRuntimeInstances
      summary: List reconciled Instances for a Runtime
      tags: [Instances]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The reconciled Instance inventory. A failed Hermes query returns cached rows with freshness UNKNOWN.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "409":
          $ref: "#/components/responses/Conflict"
    options:
      operationId: optionsRuntimeInstances
      summary: Validate CORS access for Runtime Instances
      tags: [Instances]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/instances/{instanceId}:
    parameters:
      - name: instanceId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getInstance
      summary: Read one cached Instance
      tags: [Instances]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The cached Instance record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsInstance
      summary: Validate CORS access for one Instance
      tags: [Instances]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/instances/{instanceId}/start:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    post:
      operationId: startInstance
      summary: Start an Instance
      tags: [Instances]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unused. Phase 4 always returns CAPABILITY_NOT_SUPPORTED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsStartInstance
      summary: Validate CORS access for Instance start
      tags: [Instances]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/instances/{instanceId}/stop:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    post:
      operationId: stopInstance
      summary: Stop an Instance
      tags: [Instances]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unused. Phase 4 always returns CAPABILITY_NOT_SUPPORTED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsStopInstance
      summary: Validate CORS access for Instance stop
      tags: [Instances]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/instances/{instanceId}/restart:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    post:
      operationId: restartInstance
      summary: Restart an Instance
      tags: [Instances]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unused. Phase 4 always returns CAPABILITY_NOT_SUPPORTED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    options:
      operationId: optionsRestartInstance
      summary: Validate CORS access for Instance restart
      tags: [Instances]
      security: []
      responses:
        "204":
          $ref: "#/components/responses/PreflightAccepted"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  responses:
    Unauthorized:
      description: The local session credential is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The request Origin is not an approved Desktop origin.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested local API resource does not exist. Unmatched routes use this response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: The resource does not support the request method.
      headers:
        Allow:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreflightAccepted:
      description: The known endpoint accepts the approved Desktop origin and requested headers.
    InternalError:
      description: Runtime discovery failed unexpectedly without exposing internal command output.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: The request was syntactically or semantically invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request conflicts with current Operation or installation state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[\\x21-\\x7E]+$"
    InstanceId:
      name: instanceId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Node:
      type: object
      additionalProperties: false
      required: [id, name, hostname, platform, architecture, nodeVersion, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        hostname:
          type: string
        platform:
          type: string
        architecture:
          type: string
        nodeVersion:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RuntimeDiscovery:
      type: object
      additionalProperties: false
      required: [runtimeKind, state, errorCode, selected, candidates, warnings, detectedAt, supportedRange]
      properties:
        runtimeKind:
          type: string
          enum: [hermes]
        state:
          $ref: "#/components/schemas/RuntimeDiscoveryState"
        errorCode:
          $ref: "#/components/schemas/RuntimeDiscoveryErrorCode"
        selected:
          anyOf:
            - $ref: "#/components/schemas/RuntimeCandidate"
            - type: "null"
        candidates:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/RuntimeCandidate"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/RuntimeWarning"
        detectedAt:
          type: string
          format: date-time
        supportedRange:
          type: string
...
    RuntimeInstallRequest:
      type: object
      additionalProperties: false
      properties: {}
...
    HermesPrerequisites:
      type: object
      additionalProperties: false
      required: [node, npm, nodeDependencies, checkedAt, activeOperationId]
      properties:
        node:
          $ref: "#/components/schemas/HermesPrerequisiteComponent"
        npm:
          $ref: "#/components/schemas/HermesPrerequisiteComponent"
        nodeDependencies:
          $ref: "#/components/schemas/HermesPrerequisiteComponent"
        checkedAt:
          type: string
          format: date-time
        activeOperationId:
          type: ["string", "null"]
    Operation:
      type: object
      additionalProperties: false
      required: [id, type, targetType, targetId, status, stage, progress, message, errorCode, retryable, correlationId, createdAt, startedAt, completedAt, updatedAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [runtime.install, hermes.prerequisites]
        targetType:
          type: string
        targetId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        stage:
          type: string
        progress:
          type: ["integer", "null"]
        message:
          type: string
        errorCode:
          type: ["string", "null"]
        retryable:
          type: boolean
        correlationId:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        completedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        updatedAt:
          type: string
          format: date-time
    OperationLog:
      type: object
      additionalProperties: false
      required: [operationId, correlationId, text]
      properties:
        operationId:
          type: string
        correlationId:
          type: string
        text:
          type: string
    OperationList:
      type: object
      additionalProperties: false
      required: [operations]
      properties:
        operations:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/Operation"
...